requirements.txt
tensorflow-gpu
networkx
hashlib
flufl.lock